requirements.txt
ccxt==4.*
numpy==1.26.4
pandas==2.*
pyyaml==6.*
python-dotenv==1.*
requests==2.*
